feat: add options to write `.env` and `git-hash.txt`
