+++ package.json
@@ -15,7 +15,9 @@
     "@actions/artifact": "^0.3.2",
     "@actions/core": "^1.2.4",
     "@actions/github": "^2.2.0",
-    "@actions/glob": "^0.1.0"
+    "@actions/glob": "^0.1.0",
+    "@types/fs-extra": "^9.0.1",
+    "fs-extra": "^9.0.0"
   },
   "devDependencies": {
     "@spartanbio/eslint-config": "^1.1.1",
